+++ lib/application/impl/task/task_delete_service_impl.dart
@@ -1,18 +1,24 @@
+import 'package:nova_chrono/application/api/exception/task_not_found_exception.dart';
 import 'package:nova_chrono/application/api/task/task_delete_service.dart';
 
 import '../../../domain/repository/task_repository.dart';
-import '../../../main.dart';
+import '../../../injection_container.dart';
 
 class TaskDeleteServiceImpl implements TaskDeleteService {
   late TaskRepository _taskRepository;
 
-  TaskDeleteServiceImpl({TaskRepository? taskRepository}) {
-    _taskRepository = taskRepository ?? getIt<TaskRepository>();
+  TaskDeleteServiceImpl() {
+    _taskRepository = getIt<TaskRepository>();
   }
 
   @override
   Future<void> deleteTask(String taskId) async {
-    // TODO: Check if task exists
+    var task = await _taskRepository.getById(taskId);
+
+    if (task == null) {
+      throw TaskNotFoundException('Task with id $taskId not found.');
+    }
+
     await _taskRepository.deleteTask(taskId);
   }
 }
